refactor: deleted markdown module and integrated to posts
