--- src/markdown/mod.rs
@@ -1,4 +1,0 @@
-pub mod dump;
-pub mod extract_text;
-pub mod frontmatter;
-pub mod template;
