+++ src/posts/extract_text.rs
@@ -18,7 +18,7 @@ pub fn extract_text(markdown_text: &str) -> String {
 mod test {
     use super::*;
     use crate::io::read_string;
-    use crate::markdown::frontmatter::split_frontmatter_and_content;
+    use crate::posts::frontmatter::split_frontmatter_and_content;
 
     #[test]
     fn test_extract_text() {
