+++ src/posts/index.rs
@@ -3,7 +3,7 @@ use tantivy::Index;
 
 use super::reader::get_all_posts;
 use crate::io;
-use crate::markdown::dump::dump_doc;
+use crate::posts::dump::dump_doc;
 use crate::text_engine::query::put;
 
 pub fn build(glob_pattern: &str, index: &Index) -> Result<()> {
